fix: 🐛 deterministic signature order for anonymous ABI decoding
Added deterministic order to anonymous signature to avoid inconsisencies
when decoding ABIs with them. Added mechanism to check the expected
input size of a function against the size of the data in a transaction
to find the best matching ABI instead of blindly attempting to decode in
a less predictable manner.