fix: 🐛 properly decode tuple values in call/event inputs
Restored the ability to decode tuple values in inputs of contract calls
and events. This change also adds an additional flag to the ABI
configuration that will cause ethlogger to attempt to reconcile the
shape of structs from tuples if the information in the ABI definition is
sufficient to do so (this is disabled by default).