feat: 🎸 enable reconciliation of struct data by default
This changes the default behavior to attempt to reconcile the shape of
structs from tuples in ABI information by default. If names are not
provided in the ABI definition for tuple components, then reconciliation
will fail gracfully.

BREAKING CHANGE: Output of decoded transactions and events may look
different with the change of the default behavior. Set
`reconcileStructShapeFromTuples` to false in ethlogger.yaml to restore
the behavior of previous versions.